requirements.txt
packaging==23.2
pdf2image==1.16.3
Pillow==10.1.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyPDF2==3.0.1
pytesseract==0.3.10
PyYAML==6.0.1
